Fixed UserProfileController's add/remove users to/from staff, Department Controller's UI were improved

+++ ScientificReport/Controllers/DepartmentController.cs
@@ -79,18 +79,24 @@ namespace ScientificReport.Controllers
 		// GET: Department/Create
 		[HttpGet]
 		[Authorize(Roles = UserProfileRole.Administrator)]
-		public IActionResult Create()
+		public async Task<IActionResult> Create()
 		{
 			var allUsers = _userProfileService.GetAll();
 			var allDepartments = _departmentService.GetAll();
 
-			var availableUsers = allUsers
-				.Where(user => allDepartments.All(d => d.Head.Id != user.Id && !d.Staff.Contains(user)))
-				.Select(user => new SelectItem(user.FullName, user.Id.ToString()));
+			var usersToCheck = allUsers.Where(user => allDepartments.All(d => d.Head.Id != user.Id && !d.Staff.Contains(user))).ToList();
+			var availableUsers = new List<UserProfile>();
+			foreach (var user in usersToCheck)
+			{
+				if (await _userProfileService.IsTeacherOnlyAsync(user, _userManager))
+				{
+					availableUsers.Add(user);
+				}
+			}
 			
 			return View(new DepartmentCreateModel
 			{
-				UserSelection = availableUsers
+				UserSelection = availableUsers.Select(user => new SelectItem(user.FullName, user.Id.ToString()))
 			});
 		}
 
@@ -112,15 +118,26 @@ namespace ScientificReport.Controllers
 			if (_userProfileService.UserExists(model.SelectedHeadId))
 			{
 				var head = _userProfileService.GetById(model.SelectedHeadId);
-				_departmentService.CreateItem(new Department
+				var isAvailable =
+					!(await _userProfileService.IsInRoleAsync(head, UserProfileRole.Administrator, _userManager)
+					&& await _userProfileService.IsInRoleAsync(head, UserProfileRole.HeadOfDepartment, _userManager));
+
+				if (isAvailable)
 				{
-					Title = model.Title,
-					Head = head,
-					Staff = new List<UserProfile> {head}
-				});
-				head.Position = "HeadOfDepartment";
-				_userProfileService.UpdateItem(head);
-				await _userProfileService.AddToRoleAsync(head, UserProfileRole.HeadOfDepartment, _userManager);
+					_departmentService.CreateItem(new Department
+					{
+						Title = model.Title,
+						Head = head,
+						Staff = new List<UserProfile> {head}
+					});
+					head.Position = "HeadOfDepartment";
+					_userProfileService.UpdateItem(head);
+					await _userProfileService.AddToRoleAsync(head, UserProfileRole.HeadOfDepartment, _userManager);	
+				}
+				else
+				{
+					return BadRequest();
+				}
 			}
 			else
 			{
@@ -133,7 +150,7 @@ namespace ScientificReport.Controllers
 		// GET: Department/Edit/{id}
 		[HttpGet]
 		[Authorize(Roles = UserProfileRole.HeadOfDepartmentOrAdmin)]
-		public IActionResult Edit(Guid? id)
+		public async Task<IActionResult> Edit(Guid? id)
 		{
 			if (id == null)
 			{
@@ -141,34 +158,37 @@ namespace ScientificReport.Controllers
 			}
 
 			var department = _departmentService.GetById(id.Value);
+			if (department == null)
+			{
+				return RedirectToAction("Index");
+			}
+
+			if (!IsValidCurrentUser(department))
+			{
+				return Forbid();
+			}
+
 			var departments = _departmentService.GetAllWhere(d => d.Id != department.Id).ToList();
 			var allUsers = _userProfileService.GetAll();
 
-			var availableUsers = allUsers
-				.Where(user => departments.All(d => !d.Staff.Contains(user)))
-				.Select(user => new SelectItem(user.FullName, user.Id.ToString(), user.Id.Equals(department.Head.Id)));
+			var availableUsers = allUsers.Where(user => departments.All(d => !d.Staff.Contains(user)));
 
 			var availableScientificWorks = _scientificWorkService
 				.GetAllWhere(sw => departments.All(d => !d.ScientificWorks.Contains(sw)))
 				.Select(scientificWork => new SelectItem(scientificWork.Title, scientificWork.Id.ToString()));
 
-			if (department != null)
+			return View(new DepartmentEditModel
 			{
-				return View(new DepartmentEditModel
-				{
-					DepartmentId = department.Id,
-					Title = department.Title,
-					UserSelection = availableUsers,
-					Head = department.Head,
-					SelectedHeadId = department.Head.Id,
-					Staff = department.Staff,
-					ScientificWorkItems = availableScientificWorks,
-					ScientificWorks = department.ScientificWorks,
-					IsEditingByHead = department.Head.Id == _userProfileService.Get(u => u.UserName == User.Identity.Name).Id
-				});
-			}
-
-			return RedirectToAction("Index");
+				DepartmentId = department.Id,
+				Title = department.Title,
+				UserSelection = availableUsers.Select(user => new SelectItem(user.FullName, user.Id.ToString(), user.Id.Equals(department.Head.Id))),
+				Head = department.Head,
+				SelectedHeadId = department.Head.Id,
+				Staff = department.Staff,
+				ScientificWorkItems = availableScientificWorks,
+				ScientificWorks = department.ScientificWorks,
+				IsEditingByHead = department.Head.Id == _userProfileService.Get(u => u.UserName == User.Identity.Name).Id
+			});
 		}
 
 		// POST: Department/Edit/{id}
@@ -187,10 +207,14 @@ namespace ScientificReport.Controllers
 			}
 
 			var department = _departmentService.GetById(id.Value);
-
 			if (department == null)
 			{
 				return NotFound();
+			}
+
+			if (!IsValidCurrentUser(department))
+			{
+				return Forbid();
 			}
 
 			department.Title = model.Title;
@@ -219,21 +243,31 @@ namespace ScientificReport.Controllers
 		// POST: Department/AddUserToStaff/{departmentId}
 		[HttpPost]
 		[Authorize(Roles = UserProfileRole.HeadOfDepartmentOrAdmin)]
-		public IActionResult AddUserToStaff(Guid? id, [FromBody] DepartmentUpdateStaffRequest request)
+		public async Task<IActionResult> AddUserToStaff(Guid? id, [FromBody] DepartmentUpdateStaffRequest request)
 		{
 			if (id == null)
 			{
 				return NotFound();
 			}
-
+			
 			var department = _departmentService.GetById(id.Value);
 			if (department == null)
+			{
 				return Json(ApiResponse.Fail);
+			}
+			
+			if (!IsValidCurrentUser(department))
+			{
+				return Json(ApiResponse.Fail);
+			}
 
 			var user = _userProfileService.GetById(request.UserId);
+			var isNotAvailable = await _userProfileService.IsInRoleAsync(user, UserProfileRole.HeadOfDepartment, _userManager);
 			
-			if (user == null || _departmentService.UserIsHired(user))
+			if (user == null || _departmentService.UserIsHired(user) || isNotAvailable)
+			{
 				return Json(ApiResponse.Fail);
+			}
 
 			department.Staff.Add(user);
 			_departmentService.UpdateItem(department);
@@ -253,9 +287,16 @@ namespace ScientificReport.Controllers
 
 			var department = _departmentService.GetById(id.Value);
 			if (department == null)
+			{
 				return Json(ApiResponse.Fail);
-			var user = _userProfileService.GetById(request.UserId);
+			}
+
+			if (!IsValidCurrentUser(department))
+			{
+				return Json(ApiResponse.Fail);
+			}
 			
+			var user = _userProfileService.GetById(request.UserId);
 			if (user == null || !department.Staff.Contains(user) || department.Head.Id.Equals(user.Id))
 			{
 				return Json(ApiResponse.Fail);
@@ -279,10 +320,20 @@ namespace ScientificReport.Controllers
 
 			var department = _departmentService.GetById(id.Value);
 			if (department == null)
+			{
 				return Json(new
 				{
 					Success = false
 				});
+			}
+
+			if (!IsValidCurrentUser(department))
+			{
+				return Json(new
+				{
+					Success = false
+				});
+			}
 			
 			var scientificWork = _scientificWorkService.GetById(request.ScientificWorkId);
 			if (scientificWork == null ||
@@ -315,40 +366,63 @@ namespace ScientificReport.Controllers
 
 			var department = _departmentService.GetById(id.Value);
 			if (department == null)
-				return Json(new
-				{
-					Success = false
-				});
+			{
+				return Json(ApiResponse.Fail);
+			}
+
+			if (!IsValidCurrentUser(department))
+			{
+				return Json(ApiResponse.Fail);
+			}
+			
 			var scientificWork = _scientificWorkService.GetById(request.ScientificWorkId);
 			if (scientificWork == null || !department.ScientificWorks.Contains(scientificWork))
 			{
-				return Json(new
-				{
-					Success = false
-				});
+				return Json(ApiResponse.Fail);
 			}
 
 			department.ScientificWorks.Remove(scientificWork);
 			_departmentService.UpdateItem(department);
 
-			return Json(new
-			{
-				Success = true
-			});
+			return Json(ApiResponse.Ok);
 		}
 
 		// POST: Department/Delete/{id}
 		[HttpPost]
 		[Authorize(Roles = UserProfileRole.Administrator)]
-		public IActionResult Delete(Guid? id)
+		public async Task<IActionResult> Delete(Guid? id)
 		{
 			if (id == null || !_departmentService.Exists(id.Value))
 			{
 				return NotFound();
 			}
 
-			_departmentService.DeleteById(id.Value);
+			var departmentToDelete = _departmentService.GetById(id.Value);
+			if (!IsValidCurrentUser(departmentToDelete))
+			{
+				return Forbid();
+			}
+
+			if (departmentToDelete.Staff.Count == 1)
+			{
+				await _userProfileService.RemoveFromRoleAsync(departmentToDelete.Head, UserProfileRole.HeadOfDepartment, _userManager);
+				departmentToDelete.Head.Position = UserProfileRole.Teacher;
+				_userProfileService.UpdateItem(departmentToDelete.Head);
+				_departmentService.DeleteById(id.Value);
+			}
+
 			return RedirectToAction("Index");
+		}
+
+		private bool IsValidCurrentUser(Department department)
+		{
+			var currentUser = _userProfileService.Get(User);
+			if (!PageHelpers.IsAdmin(User))
+			{
+				return currentUser.Id == department.Head.Id;	
+			}
+
+			return true;
 		}
 	}
 }
